ints must be at least size_t

+++ 0x1A-sorting_algorithms/2-selection_sort.c
@@ -9,10 +9,10 @@
 
 void selection_sort(int *array, size_t size)
 {
-	int i;
-	int j;
-	int min_i;
-	int temp;
+	size_t i;
+	size_t j;
+	size_t min_i;
+	size_t temp;
 
 	for (i = 0; i < size - 1; i++)
 	{
